+++ frontend/components/Report/View/ReportView.tsx
@@ -1,4 +1,4 @@
-import Report from '@/models/Report'
+import Report, { parseReport } from '@/models/Report'
 import React from 'react'
 import styled from 'styled-components'
 import UiGrid from '@/components/Ui/Grid/UiGrid'
@@ -8,7 +8,7 @@ import UiTextWithIcon from '@/components/Ui/TextWithIcon/UiTextWithIcon'
 import UiIcon from '@/components/Ui/Icon/UiIcon'
 import TaskList from '@/components/Task/List/TaskList'
 import { useTasksOfReport } from '@/stores/TaskStore'
-import BackendService from '@/services/BackendService'
+import BackendService, { BackendResponse } from '@/services/BackendService'
 import ReportStore from '@/stores/ReportStore'
 import UiIconButtonGroup from '@/components/Ui/Icon/Button/Group/UiIconButtonGroup'
 import UiIconButton from '@/components/Ui/Icon/Button/UiIconButton'
@@ -16,6 +16,7 @@ import { useUser } from '@/stores/UserStore'
 import UiModal from '@/components/Ui/Modal/UiModal'
 import { useIncident } from '@/stores/IncidentStore'
 import ReportForm from '@/components/Report/Form/ReportForm'
+import UiButton from '@/components/Ui/Button/UiButton'
 
 interface Props {
   report: Report
@@ -27,16 +28,44 @@ const ReportView: React.VFC<Props> = ({ report }) => {
 
   const tasks = useTasksOfReport(report.id)
 
+  const handleClose = async () => {
+    if (report.isDone) {
+      alert('Es sind alle Aufträge geschlossen.')
+    } else {
+      if (confirm(`Sind sie sicher, dass sie die Meldung "${report.title}" schliessen wollen?`)) {
+        const newReport = { ...report, isClosed: true }
+        const [data, error]: BackendResponse<Report> = await BackendService.update(`incidents/${report.incidentId}/reports`, report.id, newReport)
+        if (error !== null) {
+          throw error
+        }
+        ReportStore.save(parseReport(data))
+      }
+    }
+  }
+
+  const handleReopen = async () => {
+    if (report.isDone) {
+      alert('Es sind alle Aufträge geschlossen.')
+    } else {
+      if (confirm(`Sind sie sicher, dass sie die Meldung "${report.title}" öffnen wollen?`)) {
+        const newReport = { ...report, isClosed: false }
+        const [data, error]: BackendResponse<Report> = await BackendService.update(`incidents/${report.incidentId}/reports`, report.id, newReport)
+        if (error !== null) {
+          throw error
+        }
+        ReportStore.save(parseReport(data))
+      }
+    }
+  }
+
   const handleDelete = async () => {
     if (confirm(`Sind sie sicher, dass sie die Meldung "${report.title}" löschen wollen?`)) {
       await BackendService.delete(`incidents/${report.incidentId}/reports`, report.id)
-
       ReportStore.remove(report.id)
     }
   }
 
   const startDate = report.startsAt !== null ? report.startsAt : report.createdAt
-
   const incident = useIncident(report.incidentId)
   if (incident === null) {
     throw new Error('incident of report not found')
@@ -53,6 +82,18 @@ const ReportView: React.VFC<Props> = ({ report }) => {
         <HorizontalSpacer>
           <UiDateLabel start={startDate} end={report.endsAt} />
           <UiIconButtonGroup>
+            {/*TODO add close and reopen icon*/}
+            {!report.isDone && (
+              report.isClosed ? (
+                <UiButton onClick={handleReopen}>
+                  Öffnen
+                </UiButton>
+              ) : (
+                <UiButton onClick={handleClose}>
+                  Schliessen
+                </UiButton>
+              )
+            )}
             <UiIconButton onClick={() => alert('not yet implemented')}>
               <UiIcon.PrintAction />
             </UiIconButton>