+++ frontend/components/Ui/Header/UiHeader.tsx
@@ -1,7 +1,5 @@
 import React from 'react'
 import styled from 'styled-components'
-import Image from 'next/image'
-import UiButton from '@/components/Ui/Button/UiButton'
 import SessionStore, { useSession } from '@/stores/SessionStore'
 import { useRouter } from 'next/router'
 import UiLink from '@/components/Ui/Link/UiLink'
@@ -25,62 +23,60 @@ const UiHeader: React.VFC = () => {
   }
 
   return (
-    <HeaderContainer>
+    <Header>
       <NavContainer>
         <ImageContainer>
           <UiLink href="/">
-            <Image src="/RFOBaden_Logo_RGB.svg" alt="RFO Baden Logo" width="150" height="21" />
+            {/* eslint-disable-next-line @next/next/no-img-element */}
+            <img src="/RFOBaden_Logo_RGB.svg" alt="RFO Baden Logo" width="150" height="21" />
           </UiLink>
         </ImageContainer>
         <nav>
           <NavBar>
-            <UiHeaderItem href="/benutzer">
-              Benutzer
-            </UiHeaderItem>
             <UiHeaderItem href="/ereignisse">
               Ereignisse
+            </UiHeaderItem>
+            <UiHeaderItem href="/benutzer">
+              Benutzer
             </UiHeaderItem>
           </NavBar>
         </nav>
       </NavContainer>
       <ButtonList>
-        <UiLink href="/changelog">
-          <UiButton>
-            <UiIcon.Clipboard />
-          </UiButton>
-        </UiLink>
+        <UiHeaderItem href="/changelog" title="Changelog">
+          <UiIcon.Changelog />
+        </UiHeaderItem>
         {currentUser === null ? (
-          <UiLink href="/anmelden">
-            <UiButton type="button">
-              → anmelden
-            </UiButton>
-          </UiLink>
+          <UiHeaderItem href="/anmelden">
+            <UiIcon.Login />
+            <span>anmelden</span>
+          </UiHeaderItem>
         ) : (
-          <React.Fragment>
-            <UiLink href="/profil">
-              <UiButton type="button">
-                {currentUser.firstName} {currentUser.lastName}
-              </UiButton>
-            </UiLink>
-            <UiButton onClick={logout}>
-              abmelden →
-            </UiButton>
-          </React.Fragment>
+          <ButtonList isNarrow>
+            <UiHeaderItem href="/profil" title="Profil">
+              <UiIcon.UserInCircle />
+            </UiHeaderItem>
+            <UiHeaderItem onClick={logout}>
+              <span>{currentUser.firstName} {currentUser.lastName}</span>
+              <UiIcon.Logout />
+            </UiHeaderItem>
+          </ButtonList>
         )}
       </ButtonList>
-    </HeaderContainer>
+    </Header>
   )
 }
 export default UiHeader
 
 
-const HeaderContainer = styled.header`
+const Header = styled.header`
   display: flex;
   justify-content: space-between;
   width: 100%;
   height: 4rem;
   padding: 10px 50px 10px 50px;
   margin-bottom: 3rem;
+  color: ${({ theme }) => theme.colors.secondary.contrast};
   background: ${({ theme }) => theme.colors.secondary.value};
 `
 const NavContainer = styled.div`
@@ -89,16 +85,25 @@ const NavContainer = styled.div`
 const ImageContainer = styled.div`
   display: flex;
   align-items: center;
+  
+  img {
+    transition: 150ms ease;
+    transition-property: transform;
+    
+    :hover {
+      transform: scale(1.05);
+    }
+  }
 `
 const NavBar = styled.ul`
   display: flex;
-  gap: 1rem;
+  gap: 2rem;
   align-items: center;
   margin-left: 2rem;
   list-style: none;
 `
-const ButtonList = styled.div`
+const ButtonList = styled.div<{ isNarrow?: boolean }>`
   display: flex;
-  gap: 1rem;
+  gap: ${({ isNarrow }) => isNarrow ? '0.75rem' : '2rem'};
   align-items: center;
 `